+++ components/itemActions.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 
-export default class AppBar extends React.Component {
+import ItemButton from '../components/itemButton';
+
+export default class ItemActions extends React.Component {
   render() {}
 }