+++ components/itemTitle.jsx
@@ -1,5 +1,11 @@
 import React from 'react';
 
-export default class AppBar extends React.Component {
-  render() {}
+export default class ItemTitle extends React.Component {
+  render() {
+    return (
+      <h1>
+        <strong>{this.props.title}</strong>
+      </h1>
+    );
+  }
 }