+++ next.config.js
@@ -1,15 +1,11 @@
-// next.config.js is not transformed by Babel. So you can only use javascript features supported by your version of Node.js.
 const withCss = require('@zeit/next-css');
+const withPlugins = require('next-compose-plugins');
 
-module.exports = withCss({
+const nextConfig = {
   webpack: (config, { buildId, dev, isServer, defaultLoaders }) => {
-    // Perform customizations to webpack config
-    // Important: return the modified config
     return config;
   },
   webpackDevMiddleware: config => {
-    // Perform customizations to webpack dev middleware config
-    // Important: return the modified config
     return config;
   },
   publicRuntimeConfig: {
@@ -17,4 +13,6 @@ module.exports = withCss({
       ? 'https://case-study.anthony.codes'
       : 'http://localhost:4242',
   },
-});
+};
+
+module.exports = withPlugins([withCss], nextConfig);